# Fixes #5844 - stringify removes nested empty values