### Performance results

Methodology:

- 1,000,000 iterations of JSON.stringify on an object with a nested array.
- Run 3 times
- Take the best of 3

| Method                            | Time   | Relative to JSON.stringify | Relative to old stringify |
| --------------------------------- | ------ | -------------------------- | ------------------------- |
| JSON.stringify                    | 275 ms | 1.0x                       | 0.3x                      |
| Old stringify                     | 850 ms | 3.1x                       | 1.0x                      |
| New stringify                     | 597 ms | 2.2x                       | 0.7x                      |
| New stringify with ownKey guard   | 716 ms | 2.6x                       | 0.8x                      |
| New stringify with Object.entries | 910 ms | 3.3x                       | 1.1x                      |