@codyebberson 2025-02-20
Oo, this is a sneaky edge case.  Glad we caught it, even if indirectly.

![image](https://github.com/user-attachments/assets/f178e86f-2ad1-4128-9a6e-5a588086f038)